requirements/production.txt
-r base.txt

# Security
bleach
#cef

# Logging
#raven

# Celery: Message queue
celery
django-celery

# L10n/i18n
#Babel
#-e git://github.com/clouserw/tower.git#egg=tower

# Caching
python-memcached

# Admin
django-debug-toolbar

# Migrations
South

# API
djangorestframework
markdown
django-filter

# Server
gunicorn
